refactor : 패키지 의존관계 재설정
