--- src/main/java/baseball/domain/object/BallContainer.java
@@ -1,11 +1,0 @@
-package baseball.domain.object;
-
-public class BallContainer {
-    final Ball pitchBall;
-    final Ball swingBall;
-
-    public BallContainer(Ball pitchBall, Ball swingBall) {
-        this.pitchBall = pitchBall;
-        this.swingBall = swingBall;
-    }
-}
