--- src/main/java/baseball/domain/object/BallPiece.java
@@ -1,11 +1,0 @@
-package baseball.domain.object;
-
-public class BallPiece {
-    final int number;
-    final int position;
-
-    public BallPiece(int number, int position) {
-        this.number = number;
-        this.position = position;
-    }
-}
