+++ src/main/java/baseball/domain/object/Hitter.java
@@ -1,7 +1,7 @@
 package baseball.domain.object;
 
-import baseball.domain.object.Ball;
-import baseball.domain.object.BallContainer;
+import baseball.domain.ball.Ball;
+import baseball.domain.ball.BallContainer;
 
 public interface Hitter {
     BallContainer swing(Ball pitchBall);
