+++ src/main/java/baseball/domain/object/Pitcher.java
@@ -1,6 +1,6 @@
 package baseball.domain.object;
 
-import baseball.domain.object.Ball;
+import baseball.domain.ball.Ball;
 
 public interface Pitcher {
     Ball pitch();
