+++ src/main/java/baseball/domain/object/RefereeImpl.java
@@ -1,5 +1,8 @@
 package baseball.domain.object;
 
+import baseball.domain.ball.Ball;
+import baseball.domain.ball.BallContainer;
+import baseball.domain.ball.BallPiece;
 import baseball.domain.enums.PitchResult;
 import baseball.domain.score.Score;
 
@@ -8,8 +11,8 @@ import static baseball.domain.enums.PitchResult.*;
 public class RefereeImpl implements Referee {
     @Override
     public Score calcScore(BallContainer ballContainer) {
-        Ball swingBall = ballContainer.swingBall;
-        Ball pitchBall = ballContainer.pitchBall;
+        Ball swingBall = ballContainer.getSwingBall();
+        Ball pitchBall = ballContainer.getPitchBall();
 
         Score score = new Score();
 
@@ -22,13 +25,13 @@ public class RefereeImpl implements Referee {
     }
 
     private PitchResult checkPitchResult(Ball swingBall, BallPiece pitchBallPiece) {
-        BallPiece ballPiece = swingBall.getPieceByNumber(pitchBallPiece.number);
+        BallPiece ballPiece = swingBall.getPieceByNumber(pitchBallPiece.getNumber());
 
         if (ballPiece == null) {
             return NOTHING;
         }
 
-        if (ballPiece.position == pitchBallPiece.position) {
+        if (ballPiece.getPosition() == pitchBallPiece.getPosition()) {
             return STRIKE;
         }
